Fix issue caused by sanity by setting home to landing page
