--- pages/[slug].js
@@ -1,41 +1,0 @@
-import Layout from '../components/Layout'
-import InfoPage from '../components/pageTypes/InfoPage'
-import LandingPage from '../components/pageTypes/LandingPage'
-import { getPageBySlug, getPagePaths } from '../sanity/lib/client'
-
-const pageTypes = {
-  infoPage: InfoPage,
-  landingPage: LandingPage,
-}
-
-export default function DynamicPage({ page }) {
-  const PageType = pageTypes[page._type]
-  return <PageType content={page} />
-}
-
-export async function getStaticProps(ctx) {
-  const { slug = '' } = ctx.params
-
-  const page = await getPageBySlug({ slug })
-
-  if (!page) {
-    return {
-      notFound: true,
-    }
-  }
-
-  return {
-    props: {
-      page,
-    },
-  }
-}
-
-export async function getStaticPaths() {
-  const paths = await getPagePaths()
-
-  return {
-    paths: paths?.map((slug) => `/${slug}`) || [],
-    fallback: false,
-  }
-}
